requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
langgraph
chromadb
pypdf
python-dotenv
streamlit
sentence-transformers